requirements.txt
Flask==2.2.0
mastodon.py==1.8.0
beautifulsoup4==4.12.0
openai==0.27.2
selenium==4.8.2
pytest==7.2.2